+++ mobile/build.gradle
@@ -6,8 +6,8 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "de.rhaeus.dndsync"
-        minSdk 24
+        applicationId "it.silleellie.dndsync"
+        minSdk 26
         targetSdk 31
         versionCode 4
         versionName "2.2"
