+++ mobile/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.DNDSync">
         <activity
-            android:name=".MainActivity"
+            android:name="it.silleellie.dndsync.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -20,7 +20,7 @@
         </activity>
 
         <service
-            android:name=".DNDNotificationService"
+            android:name="it.silleellie.dndsync.DNDNotificationService"
             android:exported="true"
             android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE">
             <intent-filter>
@@ -28,7 +28,7 @@
             </intent-filter>
         </service>
         <service
-            android:name=".DNDSyncListenerService"
+            android:name="it.silleellie.dndsync.DNDSyncListenerService"
             android:enabled="true"
             android:exported="true">
             <intent-filter>
