+++ mobile/src/main/java/it/silleellie/dndsync/MainActivity.java
@@ -1,4 +1,4 @@
-package de.rhaeus.dndsync;
+package it.silleellie.dndsync;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
