+++ mobile/src/main/res/layout/activity_main.xml
@@ -5,4 +5,4 @@
     android:id="@+id/fragment_container_view"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:name="de.rhaeus.dndsync.MainFragment" />
+    android:name="it.silleellie.dndsync.MainFragment" />
