+++ wear/src/main/java/it/silleellie/dndsync/MainFragment.java
@@ -1,4 +1,4 @@
-package de.rhaeus.dndsync;
+package it.silleellie.dndsync;
 
 import android.Manifest;
 import android.app.NotificationManager;
